retriever/requirements.txt
# Core dependencies for embeddings and search
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.0.1

# Additional dependencies for RAG indexer
tqdm==4.66.1  # Progress bars
python-magic==0.4.27  # Optional: file type detection
python-magic-bin==0.4.14  # Windows support for python-magic

# Database for metadata storage
# SQLite3 is included in Python standard library